requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
google_cloud==0.34.0
google_cloud_speech==2.25.1
google_cloud_language==2.13.3